Broke out tests to easier test one at the time
